rebase2, dpk not being read
