+++ app/src/state/actions/general-actions.ts
@@ -127,7 +127,6 @@ export const fetchDatapackIndex = action("fetchDatapackIndex", async () => {
     setDatapackIndex({ ...state.datapackIndex, ...datapackIndex });
     console.log("Datapacks loaded");
   } catch (e) {
-    console.log("Error in fetchDatapackIndex", e);
     displayServerError(null, ErrorCodes.SERVER_RESPONSE_ERROR, ErrorMessages[ErrorCodes.SERVER_RESPONSE_ERROR]);
     console.error(e);
   }
@@ -160,7 +159,6 @@ export const fetchMapPackIndex = action("fetchMapPackIndex", async () => {
     setMapPackIndex({ ...state.mapPackIndex, ...mapPackIndex });
     console.log("MapPacks loaded");
   } catch (e) {
-    console.log("Error in fetchMapPackIndex", e);
     displayServerError(null, ErrorCodes.SERVER_RESPONSE_ERROR, ErrorMessages[ErrorCodes.SERVER_RESPONSE_ERROR]);
     console.error(e);
   }
@@ -269,7 +267,6 @@ export const fetchUserDatapacks = action("fetchUserDatapacks", async () => {
       displayServerError(data, ErrorCodes.INVALID_USER_DATAPACKS, ErrorMessages[ErrorCodes.INVALID_USER_DATAPACKS]);
     }
   } catch (e) {
-    console.log("Error in fetchUserDatapacks", e);
     displayServerError(null, ErrorCodes.SERVER_RESPONSE_ERROR, ErrorMessages[ErrorCodes.SERVER_RESPONSE_ERROR]);
     console.error(e);
   }
@@ -285,7 +282,7 @@ export const uploadDatapack = action("uploadDatapack", async (file: File, metada
   formData.append("file", file);
   formData.append("title", title);
   formData.append("description", description);
-  formData.append("isPublic", String(isPublic));
+  formData.append("isPublic", isPublic);
   formData.append("references", JSON.stringify(references));
   formData.append("tags", JSON.stringify(tags));
   formData.append("authoredBy", authoredBy);
