+++ server/__tests__/admin-routes.test.ts
@@ -907,6 +907,7 @@ describe("adminUploadServerDatapack", () => {
     title: "test-title",
     size: "30MB",
     date: "2021-01-01",
+    isPublic: "false",
     tags: ["test-tag"],
     references: ["test-reference"],
     contact: "test-contact",
@@ -1302,6 +1303,7 @@ describe("adminDeleteServerDatapack", () => {
     file: "active-datapack.dpk",
     size: "30MB",
     date: "2021-01-01",
+    isPublic: "false",
     tags: ["test-tag"],
     references: ["test-reference"],
     contact: "test-contact",
