+++ server/src/routes/user-routes.ts
@@ -4,7 +4,15 @@ import path from "path";
 import { runJavaEncrypt } from "../encryption.js";
 import { assetconfigs, checkHeader, resetUploadDirectory } from "../util.js";
 import { MultipartFile } from "@fastify/multipart";
-import { assertDatapackIndex, assertMapPackIndex, DatapackIndex, MapPackIndex } from "@tsconline/shared";
+import {
+  assertDatapackIndex,
+  assertIndexResponse,
+  assertMapPackIndex,
+  DatapackIndex,
+  DatapackParsingPack,
+  MapPack,
+  MapPackIndex
+} from "@tsconline/shared";
 import { exec } from "child_process";
 import { createWriteStream } from "fs";
 import { pipeline } from "stream/promises";
@@ -134,6 +142,54 @@ export const requestDownload = async function requestDownload(
   }
 };
 
+export const fetchPublicDatapacks = async function fetchPublicDatapacks(request: FastifyRequest, reply: FastifyReply) {
+  const datapackIndex: DatapackIndex = {};
+  const mapPackIndex: MapPackIndex = {};
+
+  try {
+    try {
+      await access(path.join(assetconfigs.publicDirectory, "DatapackIndex.json"));
+    } catch (e) {
+      reply.status(200).send({
+        datapackIndex: datapackIndex,
+        mapPackIndex: mapPackIndex
+      });
+      return;
+    }
+    try {
+      const publicDatapackPath = await readFile(path.join(assetconfigs.publicDirectory, "DatapackIndex.json"), "utf8");
+      const parsedDatapackIndex = JSON.parse(publicDatapackPath);
+      assertDatapackIndex(parsedDatapackIndex);
+      if (parsedDatapackIndex) {
+        Object.assign(datapackIndex, parsedDatapackIndex);
+      }
+    } catch (e) {
+      console.error("Error loading public datapacks:", e);
+    }
+    try {
+      const publicMapPackPath = await readFile(path.join(assetconfigs.publicDirectory, "MapPackIndex.json"), "utf8");
+      const parsedMapPackIndex = JSON.parse(publicMapPackPath);
+      assertMapPackIndex(parsedMapPackIndex);
+      if (parsedMapPackIndex) {
+        Object.assign(mapPackIndex, parsedMapPackIndex);
+      }
+    } catch (e) {
+      console.error("Error loading public mappacks:", e);
+    }
+    const indexResponse = { datapackIndex, mapPackIndex };
+    try {
+      assertIndexResponse(indexResponse);
+      reply.status(200).send(indexResponse);
+    } catch (e) {
+      console.error("Invalid index response structure:", e);
+      reply.status(500).send({ error: "Invalid index response structure" });
+    }
+  } catch (e) {
+    reply.status(500).send({ error: "Failed to load public datapacks" });
+    console.error("Failed to load public datapacks:", e);
+  }
+};
+
 export const fetchUserDatapacks = async function fetchUserDatapacks(request: FastifyRequest, reply: FastifyReply) {
   // for test usage: const uuid = "username";
   const uuid = request.session.get("uuid");
@@ -239,6 +295,7 @@ export const uploadDatapack = async function uploadDatapack(request: FastifyRequ
     reply.status(400).send({ error: "No file uploaded" });
     return;
   }
+  const isPublic = fields.isPublic === "true";
   const filename = uploadedFile.filename;
   fields.filename = filename;
   fields.filepath = filepath;
@@ -347,6 +404,54 @@ export const uploadDatapack = async function uploadDatapack(request: FastifyRequ
   } catch (e) {
     await errorHandler("Failed to load and write metadata for file", 500, e);
     return;
+  }
+  if (isPublic) {
+    const publicDatapackPath = path.join(assetconfigs.publicDirectory, "DatapackIndex.json");
+    const publicMappackPath = path.join(assetconfigs.publicDirectory, "MapPackIndex.json");
+    const publicFilesPath = path.join(assetconfigs.publicDirectory, "publicFiles.json");
+    await mkdir(assetconfigs.publicDirectory, { recursive: true });
+
+    let publicDatapacks: DatapackIndex = {};
+    let publicMappacks: MapPackIndex = {};
+    let publicFiles: string[] = [];
+
+    try {
+      const publicData = await readFile(publicDatapackPath, "utf-8");
+      publicDatapacks = JSON.parse(publicData);
+      assertDatapackIndex(publicDatapacks);
+      const publicMapData = await readFile(publicMappackPath, "utf-8");
+      publicMappacks = JSON.parse(publicMapData);
+      assertMapPackIndex(publicMappacks);
+      const publicFilesData = await readFile(publicFilesPath, "utf-8");
+      publicFiles = JSON.parse(publicFilesData);
+
+      if (publicFiles.includes(filename)) {
+        await errorHandler("File ${filename} already exists in public directory", 409);
+        return;
+      }
+      publicFiles.push(filename);
+    } catch (e) {
+      if ((e as NodeJS.ErrnoException).code != "ENOENT") {
+        await errorHandler("Failed to read public JSON files", 500, e);
+        return;
+      }
+    }
+    assertDatapackIndex(publicDatapacks);
+    assertMapPackIndex(publicMappacks);
+    if (datapackIndex[filename]) {
+      publicDatapacks[filename] = datapackIndex[filename] as DatapackParsingPack; //error without cast
+    }
+    if (mapPackIndex[filename]) {
+      publicMappacks[filename] = mapPackIndex[filename] as MapPack;
+    }
+    try {
+      await writeFile(publicDatapackPath, JSON.stringify(publicDatapacks));
+      await writeFile(publicMappackPath, JSON.stringify(publicMappacks));
+      await writeFile(publicFilesPath, JSON.stringify(publicFiles));
+    } catch (e) {
+      await errorHandler("Failed to update public index files", 500, e);
+      return;
+    }
   }
   reply.status(200).send({ message: "File uploaded" });
 };
